+++ tests.py
@@ -36,19 +36,11 @@
 
 from __future__ import absolute_import
 from __future__ import print_function
+from __future__ import unicode_literals
 
-import sys
 import unittest
 
 import publicsuffix2 as publicsuffix
-
-if sys.version < '3':
-    import codecs
-    def u(x):
-        return codecs.unicode_escape_decode(x)[0]
-else:
-    def u(x):
-        return x
 
 
 class TestPublicSuffix(unittest.TestCase):
@@ -60,7 +52,8 @@ class TestPublicSuffix(unittest.TestCase):
         assert 'com' == psl.get_sld('.com')
         assert 'com' == psl.get_sld('a.example.com')
 
-        # enable strict mode
+    def test_get_sld_from_empty_list_in_strict_mode(self):
+        psl = publicsuffix.PublicSuffixList([])
         assert None == psl.get_sld('com', strict=True)
 
     def test_get_sld_from_list(self):
@@ -88,13 +81,13 @@ class TestPublicSuffix(unittest.TestCase):
         assert 'example.com' == psl.get_sld('example.com.')
 
     def test_get_sld_from_list_with_unicode(self):
-        psl = publicsuffix.PublicSuffixList([u('\u0440\u0444')], idna=False)
-        assert u('\u0440\u0444') == psl.get_sld(u('\u0440\u0444'))
-        assert u('example.\u0440\u0444') == psl.get_sld(u('example.\u0440\u0444'))
-        assert u('example.\u0440\u0444') == psl.get_sld(u('a.example.\u0440\u0444'))
-        assert u('example.\u0440\u0444') == psl.get_sld(u('a.a.example.\u0440\u0444'))
+        psl = publicsuffix.PublicSuffixList([u'\u0440\u0444'], idna=False)
+        assert u'\u0440\u0444' == psl.get_sld(u'\u0440\u0444')
+        assert u'example.\u0440\u0444' == psl.get_sld(u'example.\u0440\u0444')
+        assert u'example.\u0440\u0444' == psl.get_sld(u'a.example.\u0440\u0444')
+        assert u'example.\u0440\u0444' == psl.get_sld(u'a.a.example.\u0440\u0444')
 
-    def test_get_public_suffix_from_builtin_full_publicsuffix_org_using_func(self):
+    def test_get_public_suffix_from_builtin_full_publicsuffix_org_using_top_level_function(self):
         assert 'com' == publicsuffix.get_public_suffix('COM')
         assert 'example.com' == publicsuffix.get_public_suffix('example.COM')
         assert 'example.com' == publicsuffix.get_public_suffix('WwW.example.COM')
@@ -192,9 +185,9 @@ class TestPublicSuffixUsingTheCurrentVendoredPSL(unittest.TestCase):
         assert 'test.k12.ak.us' == psl.get_sld('www.test.k12.ak.us')
 
 
-class TestPublicSuffixIdna(unittest.TestCase):
+class TestPublicSuffixGetSldIdna(unittest.TestCase):
 
-    def test_idna_encoded(self):
+    def test_get_sld_idna_encoded(self):
         # actually the default
         psl = publicsuffix.PublicSuffixList(idna=True)
         assert 'xn--85x722f.com.cn' == psl.get_sld('xn--85x722f.com.cn')
@@ -202,7 +195,7 @@ class TestPublicSuffixIdna(unittest.TestCase):
         assert 'xn--85x722f.xn--55qx5d.cn' == psl.get_sld('www.xn--85x722f.xn--55qx5d.cn')
         assert 'shishi.xn--55qx5d.cn' == psl.get_sld('shishi.xn--55qx5d.cn')
 
-    def test_utf8_encoded(self):
+    def test_get_sld_with_utf8_encoded(self):
         # uses the list provided utf-8 defaults
         psl = publicsuffix.PublicSuffixList(idna=False)
         assert u'食狮.com.cn' == psl.get_sld(u'食狮.com.cn')
@@ -210,13 +203,13 @@ class TestPublicSuffixIdna(unittest.TestCase):
         assert u'食狮.公司.cn' == psl.get_sld(u'www.食狮.公司.cn')
         assert u'shishi.公司.cn' == psl.get_sld(u'shishi.公司.cn')
 
-    def test_exceptions(self):
+    def test_get_sld_exceptions(self):
         psl = publicsuffix.PublicSuffixList()
         # www is the exception
         assert 'www.ck' == psl.get_sld('www.www.ck')
         assert 'this.that.ck' == psl.get_sld('this.that.ck')
 
-    def test_no_wildcard(self):
+    def test_get_sld_no_wildcard(self):
         psl = publicsuffix.PublicSuffixList()
         # test completion when no wildcards should be processed
         assert 'com.pg' == psl.get_sld('telinet.com.pg', wildcard=False)
@@ -224,70 +217,80 @@ class TestPublicSuffixIdna(unittest.TestCase):
         result = psl.get_sld('blah.ap-southeast-1.elb.amazonaws.com', wildcard=False)
         assert expected == result
 
-    def test_convenience_functions(self):
+    def test_get_sld_top_convenience_function_is_the_same_as_PublicSuffixList_method(self):
         psl = publicsuffix.PublicSuffixList()
         # these functions should be identical
-        assert psl.get_sld('www.google.com') == psl.get_sld('www.google.com')
-        assert psl.get_sld('www.test.ak.us') == psl.get_sld('www.test.ak.us')
+        assert psl.get_sld('www.google.com') == publicsuffix.get_sld('www.google.com')
+        assert psl.get_sld('www.test.ak.us') == publicsuffix.get_sld('www.test.ak.us')
 
-    def test_tld_function(self):
+    def test_get_tld_returns_correct_tld_or_etld(self):
         psl = publicsuffix.PublicSuffixList()
-        # checks that the eTLD or TLD is produced
-        assert psl.get_tld('com') == 'com'
-        self.assertEqual(psl.get_tld('city.kobe.jp'), 'kobe.jp')
-        self.assertEqual(psl.get_tld('kobe.jp'), 'kobe.jp')
-        self.assertEqual(psl.get_tld('amazonaws.com'), 'amazonaws.com')
-        assert psl.get_tld('telinet.com.pg', wildcard=True) == 'com.pg'
-        assert psl.get_tld('telinet.com.pg', wildcard=False) == 'pg'
-        self.assertEqual(psl.get_tld('com.pg', wildcard=True), 'com.pg')
-        self.assertEqual(psl.get_tld('com.pg', wildcard=False), 'pg')
-        assert psl.get_tld('telinet.co.uk', wildcard=False) == 'co.uk'
-        self.assertEqual(psl.get_tld('co.uk', wildcard=True), 'co.uk')
-        self.assertEqual(psl.get_tld('co.uk', wildcard=False), 'co.uk')
-        assert psl.get_tld('blah.local', strict=True) is None
-        self.assertEqual(psl.get_tld('blah.local', wildcard=False), None)
-        self.assertEqual(psl.get_tld('blah.local'), 'local')
+        assert 'com' == psl.get_tld('com')
+        assert 'kobe.jp' == psl.get_tld('city.kobe.jp')
+        assert 'kobe.jp' == psl.get_tld('kobe.jp')
+        assert 'amazonaws.com' == psl.get_tld('amazonaws.com')
+        assert 'com.pg' == psl.get_tld('telinet.com.pg', wildcard=True)
+        assert 'pg' == psl.get_tld('telinet.com.pg', wildcard=False)
+        assert 'com.pg' == psl.get_tld('com.pg', wildcard=True)
+        assert 'pg' == psl.get_tld('com.pg', wildcard=False)
+        assert 'co.uk' == psl.get_tld('telinet.co.uk', wildcard=False)
+        assert 'co.uk' == psl.get_tld('co.uk', wildcard=True)
+        assert 'co.uk' == psl.get_tld('co.uk', wildcard=False)
+        assert None == psl.get_tld('blah.local', strict=True)
+        assert None == psl.get_tld('blah.local', wildcard=False)
+        assert 'local' == psl.get_tld('blah.local')
 
-        # FQDNs
-        self.assertEqual(psl.get_tld('www.foo.com.'), 'com')
-        self.assertEqual(psl.get_tld('.'), '')  # the root domain
+    def test_get_tld_returns_correct_tld_or_etld_for_fqdn(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert 'com' == psl.get_tld('www.foo.com.')
 
-        # the tld for empty string
-        self.assertEqual(psl.get_tld(''), None)
+    def test_get_tld_returns_correct_tld_or_etld_for_root_domain(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert '' == psl.get_tld('.')
+
+    def test_get_tld_returns_correct_tld_or_etld_for_empty_string(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert None == psl.get_tld('')
 
     def test_PublicSuffixList_tlds_is_loaded_correctly(self):
         psl = publicsuffix.PublicSuffixList()
         assert psl.tlds
 
 
-class TestPublicSuffixIssue5(unittest.TestCase):
+class TestPublicSuffixGetSld(unittest.TestCase):
 
-    def test_backward_compatibility(self):
+    def test_get_sld_backward_compatibility(self):
         psl = publicsuffix.PublicSuffixList()
-        self.assertEqual(psl.get_sld('com'), 'com')
-        self.assertEqual(psl.get_sld('foo.com'), 'foo.com')
-        self.assertEqual(psl.get_sld('foo.co.jp'), 'foo.co.jp')
-        self.assertEqual(psl.get_sld('co.jp'), 'co.jp')
-        self.assertEqual(psl.get_sld('jp'), 'jp')
+        assert 'com' == psl.get_sld('com')
+        assert 'foo.com' == psl.get_sld('foo.com')
+        assert 'foo.co.jp' == psl.get_sld('foo.co.jp')
+        assert 'co.jp' == psl.get_sld('co.jp')
+        assert 'jp' == psl.get_sld('jp')
 
-        # strict and wildcard flags
-        self.assertEqual(psl.get_sld('local'), 'local')
-        self.assertEqual(psl.get_sld('foo.local'), 'local')
-        self.assertEqual(psl.get_sld('local', strict=True), None)
-        self.assertEqual(psl.get_sld('foo.local', strict=True), None)
-        self.assertEqual(psl.get_sld('local', wildcard=False), 'local')
-        self.assertEqual(psl.get_sld('foo.local', strict=False), 'local')
+    def test_get_sld_backward_compatibility_strict_and_wildcard_flags(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert 'local' == psl.get_sld('local')
+        assert 'local' == psl.get_sld('foo.local')
+        assert None == psl.get_sld('local', strict=True)
+        assert None == psl.get_sld('foo.local', strict=True)
+        assert 'local' == psl.get_sld('local', wildcard=False)
+        assert 'local' == psl.get_sld('foo.local', strict=False)
 
-        # the sld for empty string
-        self.assertEqual(psl.get_sld(''), None)
-        self.assertEqual(psl.get_sld('', strict=True), None)
-        self.assertEqual(psl.get_sld('', wildcard=False), None)
+    def test_get_sld_backward_compatibility_sld_for_empty_string(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert None == psl.get_sld('')
+        assert None == psl.get_sld('', strict=True)
+        assert None == psl.get_sld('', wildcard=False)
 
-        # FQDNs
-        self.assertEqual(psl.get_sld('www.foo.com.'), 'foo.com')
-        self.assertEqual(psl.get_sld('.'), '')  # the root domain
-        self.assertEqual(psl.get_sld('.', strict=True), None)  # the root domain
-        self.assertEqual(psl.get_sld('.', wildcard=False), '')  # the root domain
+    def test_get_sld_backward_compatibility_sld_for_fqdn(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert 'foo.com' == psl.get_sld('www.foo.com.')
+
+    def test_get_sld_backward_compatibility_sld_for_root_domain(self):
+        psl = publicsuffix.PublicSuffixList()
+        assert '' == psl.get_sld('.')
+        assert None == psl.get_sld('.', strict=True)
+        assert '' == psl.get_sld('.', wildcard=False)
 
 
 if __name__ == '__main__':